MultiTenantRebacDatasetLifecycleMessageConsumer: extract
