+++ src/domain/auth-rebac/services/src/lib.rs
@@ -11,7 +11,9 @@
 pub use kamu_auth_rebac as domain;
 
 mod messages;
+mod multi_tenant_rebac_dataset_lifecycle_message_consumer;
 mod rebac_service_impl;
 
 pub use messages::*;
+pub use multi_tenant_rebac_dataset_lifecycle_message_consumer::*;
 pub use rebac_service_impl::*;
